Combine the two order lambdas into one
