--- server/orders-add/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "cmm-orders-add",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "aws-api-gateway-return": "file:../local_dependencies/aws-api-gateway-return"
-  },
-  "scripts": {
-    "test": "jest"
-  },
-  "devDependencies": {
-    "jest": "^22.4.0"
-  }
-}
