--- server/orders-get/index.test.js
@@ -1,3 +1,0 @@
-test('test runs', () => {
-  expect(true);
-});
