--- server/orders-get/jest.config.js
@@ -1,6 +1,0 @@
-// jest.config.js
-module.exports = {
-  verbose: true,
-  testMatch: ['**/__tests__/**/*.js?(x)', '**/?(*.)(spec|test).js?(x)'],
-  testPathIgnorePatterns: ['/node_modules/']
-};
